# http/httpguts: speed up ValidHeaderFieldName