@jub0bs 2024-04-10
Per the [contribution guide](https://go.dev/doc/contribute#commit_messages), I didn't use any markup in the description of my PR, but here are my benchmark results formatted more nicely:

```
goos: darwin
goarch: amd64
pkg: golang.org/x/net/http/httpguts
cpu: Intel(R) Core(TM) i7-6700HQ CPU @ 2.60GHz
                       │   before    │                after                │
                       │   sec/op    │   sec/op     vs base                │
IsTokenRune-8            315.2n ± 0%   316.2n ± 1%        ~ (p=0.245 n=20)
ValidHeaderFieldName-8   62.77n ± 0%   29.16n ± 0%  -53.55% (p=0.000 n=20)
geomean                  140.7n        96.02n       -31.73%

                       │    before    │                after                │
                       │     B/op     │    B/op     vs base                 │
IsTokenRune-8            0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=20) ¹
ValidHeaderFieldName-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=20) ¹
geomean                             ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                       │    before    │                after                │
                       │  allocs/op   │ allocs/op   vs base                 │
IsTokenRune-8            0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=20) ¹
ValidHeaderFieldName-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=20) ¹
geomean                             ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```

cc @neild